requirements.txt
paramiko
tqdm
python-dotenv
rich